Move scalarizer to Operators/
